// next-react/src/gql/graphql.tsx
export enum Contact_Category_Enum {
	Account = 'ACCOUNT',
	Basic = 'BASIC'
}
// ...
export enum Experience_Category_Enum {
	Education = 'EDUCATION',
	Internship = 'INTERNSHIP',
	Job = 'JOB',
	Project = 'PROJECT',
	Volunteer = 'VOLUNTEER'
}
// ...
export enum Inscription_Type_Enum {
	Inline = 'INLINE',
	Under = 'UNDER'
}
// ...
export enum Remote_Category_Enum {
	Hybrid = 'HYBRID',
	Onsite = 'ONSITE',
	Remote = 'REMOTE'
}
// ...
export enum Skill_Category_Enum {
	FrameworkOrLibrary = 'FRAMEWORK_OR_LIBRARY',
	Language = 'LANGUAGE',
	OtherSkillOrTool = 'OTHER_SKILL_OR_TOOL',
	ProgrammingLanguage = 'PROGRAMMING_LANGUAGE'
}
// ...
export type GetInfoQuery = {
	__typename?: 'query_root',
	contacts: Array<{
		__typename?: 'contacts',
		id: any,
		name: string,
		href?: string | null,
		category: Contact_Category_Enum,
		data: string,
		visible: boolean
	}>,
	experiences: Array<{
		__typename?: 'experiences',
		id: any,
		name: string,
		name_href?: string | null,
		category: Experience_Category_Enum,
		inscription?: string | null,
		inscription_type: Inscription_Type_Enum,
		inscription_href?: string | null,
		date_start: any,
		date_end?: any | null,
		location?: string | null,
		remote: Remote_Category_Enum,
		description?: string | null,
		items?: string | null,
		visible: boolean,
		experience_skills: Array<{ __typename?: 'experience_skills', skill: { __typename?: 'skills', name: string } }>
	}>,
	skills: Array<{
		__typename?: 'skills',
		id: any,
		name: string,
		href?: string | null,
		category: Skill_Category_Enum,
		description?: string | null,
		visible: boolean
	}>
};